Correctly implement unfold-episodes
Unfolding on an episode would not unfold the whole season but only
from the episode at point.
This was first fixed by auto-moving to the season header after folding on
an episode.
This broke st-cycle by entering an endless loop of going to next invisible header
then back after folding.
Now point is not moved after folding on an episode, and unfolding correctly
unfolds the whole season and point is restored to the previously hovered episode